+++ assets/fonts/__init__.py
@@ -1,0 +1,1 @@
+"""Bundled Roboto / Roboto Serif / Roboto Mono font files (installed as `sprezzature_figures_fonts`)."""